% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEET_enrichment_barplot.R
\name{count_intersection}
\alias{count_intersection}
\title{DEET_cluster_enrichment}
\usage{
count_intersection(x)
}
\arguments{
\item{enrich_list}{A list of enrichments from DEET, with each element
post-processed with the barplot enrichment function.}

\item{outname}{A character giving the title of the barplot or dotplot.}

\item{width}{The number of inches in the barplot or dotplot.}

\item{text_angle}{The angle of the enriched studies.}

\item{horizontal}{Whether the output barplot is vertical or horizontal}

\item{topn}{the top number of studies (by p-value) to be plotted.}

\item{ol_size}{the minimum number of overlapping genes (or paths) in an enriched study.}

\item{exclude_domain}{Exclude studies enriched based on DEGs, Paths, or TF if
the user happened to aggregate the results into a single DF, generally unused.}

\item{cluster_order}{Factor to group studies based on the researchers custom annotation.}

\item{dot}{logical (T/F) of whether to produce a dotplot or a barplot}

\item{colors}{Type of color pallete to input into 'scale_fill_brewer' of ggplot.}
}
\value{
A ggplot2 object (barplot or dotplot) of enrichment identified within DEET.
}
\description{
Generate barplots or dotplots from the output of DEET
}
\examples{

data(DEET_feature_extract_example_matrix)
data(DEET_feature_extract_example_response)
single1 <- DEET_feature_extract(DEET_feature_extract_example_matrix,DEET_feature_extract_example_response,"categorical")

}
\references{

}
\author{
Dustin Sokolowski, Hauyun Hou PhD
}
