% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEET_feature_extract_example_response.R
\name{DEET_feature_extract_example_response}
\alias{DEET_feature_extract_example_response}
\title{DEET_feature_extract_example_response}
\format{
Character vector giving the source (TCGA SRA, GTEx, SRA-manual) of 200 comparisons within DEET.
}
\usage{
data(DEET_feature_extract_example_response)
}
\description{
Character vector giving the source (TCGA SRA, GTEx, SRA-manual) of 200 comparisons within DEET.
Used as the input for the `response` input of `DEET_feature_extract` in the example.
For this response variable to work, the `datatype` input variable would also need
to be set to "categorical".
}
\examples{
data(DEET_feature_extract_example_response)
}
\keyword{datasets}
