% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_DEET_enrich_input.R
\name{example_DEET_enrich_input}
\alias{example_DEET_enrich_input}
\title{example_DEET_enrich_input}
\format{
A data frame with three columns. Rows are genes and it's
populated by the gene symbol, padj of gene expression, and coef (log2Fold-change).
}
\usage{
data(example_DEET_enrich_input)
}
\description{
Exon-level DEGs of HAoEC after TNFa treatment for 45 mins
from Alizada et al., 2021. Object is a data.frame with columns
"gene_symbol" "padj" and "coef", which in this case is the
log2Fold-change of differential expression.
}
\examples{
data(example_DEET_enrich_input)
}
\keyword{datasets}
