% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEET_enrich.R
\name{single_gene_set_cor_test}
\alias{single_gene_set_cor_test}
\title{single_gene_set_cor_test}
\usage{
single_gene_set_cor_test(
  index,
  DEG_processed,
  DEET_DE,
  AP_DEET_DE_sig,
  DEET_metadata
)
}
\arguments{
\item{index}{Value between 1 and length(DEET_DE).}

\item{DEG_processed}{Processed data frame consisting of the input genes.
The 3 columns are gene_symbol, padj, and coef.}

\item{DEET_DE}{List containing DEET's DE and library of studies. This
contains a subset of studies with 3 or more overlapping DEGs.}

\item{AP_DEET_DE_sig}{Data frame consisting of the enrichment of input gene
list on DEET’s studies. This contains a subset of studies with 3 or more
overlapping DEGs.}

\item{DEET_metadata}{Corresponding metadata of studies in DEET. This
contains a subset of studies with 3 or more overlapping DEGs.}
}
\value{
A list containing 2 objects.
\itemize{
\item summary_cor - Summary of the correlations.
\item matrix - Scatterplot of genes that are DE in the study.
}
}
\description{
Calculate the Pearson and Spearman correlation between the
input DEGs and an individual set of DEGs from DEET.
}
\author{
Dustin Sokolowski, Jedid Ahn
}
